Move testing SOCKS server to own crate
